requirements.txt
numpy
pyyaml
tqdm
pillow
tensorboardX
torch
torchvision

